glance/contrib/plugins/image_artifact/requirements.txt
python-glanceclient
